Firebird: publish test results to TestRails (#2384)
