+++ edge-oai-pmh/src/main/resources/firebird/edge-oai-pmh/features/instance-suppressed-with-discovery-flag-C193960.feature
@@ -1,4 +1,3 @@
-@C193960
 Feature: instance suppressed with discovery flag C193960
 
   Background:
@@ -6,7 +5,11 @@ Feature: instance suppressed with discovery flag C193960
     * callonce variables
     * callonce login testUser
     * def okapiTokenAdmin = okapitoken
+    * callonce read('init_data/update-configuration.feature@TechnicalConfig')
+    * callonce read('init_data/update-configuration.feature@BehaviorConfig')
+    * callonce read('init_data/init-edge-oai-pmh.feature')
 
+  @C193960
   Scenario: Verify suppressed records are included with discovery flag when configured to transfer suppressed records
     # Set technical config to avoid pagination
     Given path 'oai-pmh/configuration-settings'
