+++ mod-data-export/src/test/java/org/folio/ModDataExportApiTest.java
@@ -87,12 +87,6 @@ public class ModDataExportApiTest extends TestBaseEureka {
     }
 
     @Test
-    @Order(8)
-    void fileExportForMarcAuthorityRecordExportTest() {
-        runFeatureTest("export-for-authority");
-    }
-
-    @Test
     @Order(9)
     void deleteJobExecutionTest() {
         runFeatureTest("delete-job-execution");
@@ -114,48 +108,6 @@ public class ModDataExportApiTest extends TestBaseEureka {
     @Order(12)
     void dataExportDeletedMarcIdsTest() {
         runFeatureTest("export-deleted-marc-ids");
-    }
-
-    @Test
-    @Order(15)
-    void dataExportDeletedAuthoritiesTest() {
-        runFeatureTest("export-deleted-authorities");
-    }
-
-    @Test
-    @Order(16)
-    void ExportInstanceConfigC431148Test() {
-        runFeatureTest("export-for-instance-config-C431148");
-    }
-
-    @Test
-    @Order(17)
-    void ExportInstanceConfigCqlC432309Test() {
-        runFeatureTest("export-for-instance-config-cql-C432309");
-    }
-
-    @Test
-    @Order(18)
-    void ExportHoldingConfigC432311Test() {
-        runFeatureTest("export-for-holding-config-C432311");
-    }
-
-    @Test
-    @Order(19)
-    void ExportAuthorityConfigC432314Test() {
-        runFeatureTest("export-authority-config-C432314");
-    }
-
-    @Test
-    @Order(20)
-    void NegativeConfigExportC432315Test() {
-        runFeatureTest("negative-config-export-C432315");
-    }
-
-    @Test
-    @Order(21)
-    void dataExportDeletedAuthoritiesDeletedProfileTest() {
-        runFeatureTest("export-deleted-authorities-deleted-profile");
     }
 
     @Test
