+++ mod-data-export/src/test/resources/firebird/dataexport/features/export-authority-config-C432314.feature
@@ -14,6 +14,7 @@ Feature: Verify configured limit of exported file size - Authorities (UUID)
     * configure headers = headersUser
     * configure retry = { interval: 15000, count: 10 }
 
+  @C432314
   Scenario Outline: setting configuration slice_size to 1, test upload file and export flow for authority uuids when related MARC_AUTHORITY records exist.
     #should create file definition
 
@@ -123,6 +124,7 @@ Feature: Verify configured limit of exported file size - Authorities (UUID)
       | fileName                                    | uploadFormat |
       | test-export-config-authority-csv.csv        | csv          |
 
+  @C432314
   Scenario Outline: test upload file and export flow for authority uuids when related MARC_AUTHORITY records exist.
     #should create file definition
 
@@ -205,6 +207,7 @@ Feature: Verify configured limit of exported file size - Authorities (UUID)
       | fileName                                    | uploadFormat |
       | test-export-config-authority-csv.csv        | csv          |
 
+  @C432314
   Scenario Outline: test upload file and export flow for authority uuids when related MARC_AUTHORITY records exist.
     #should create file definition
