+++ mod-data-export/src/test/resources/firebird/dataexport/features/export-deleted-authorities.feature
@@ -1,4 +1,5 @@
 @parallel=false
+
 Feature: Test export authority deleted
 
   Background:
@@ -20,6 +21,7 @@ Feature: Test export authority deleted
 
     * configure retry = { interval: 15000, count: 10 }
 
+  @C566112
   Scenario: Test export deleted authority
     # create authority that will be deleted
     * def authorityDeletedId = '0410f607-60eb-4a5d-9951-28f3794941ec'
